use [u8; 6] in public interface and only use MacAddr internally
